setting up model tests
