+++ time_series/test_time_series/test_model.py
@@ -1,0 +1,10 @@
+from time_series import lotka_voltera as lv
+
+
+class TestLotkaVoltera:
+    def test_step(self, species_collection):
+        model = lv.LotkaVolteraModel(
+            species_collection
+        )
+
+        assert model.capacity == 1.0
