+++ time_series/time_series/lotka_voltera.py
@@ -1,0 +1,11 @@
+from time_series.species import SpeciesCollection
+
+
+class LotkaVolteraModel:
+    def __init__(
+            self,
+            species_collection: SpeciesCollection,
+            capacity: float = 1.0
+    ):
+        self.species_collection = species_collection
+        self.capacity = capacity
